updated toupcam drivers and python library, and added some variable definitions to camera_toupcam that most things look for

+++ software/control/gui_toupcam_IMX571.py
@@ -192,8 +192,8 @@ class OctopiGUI(QMainWindow):
 			main_dockArea.addDock(dock_controlPanel,'right')
 			self.setCentralWidget(main_dockArea)
 			desktopWidget = QDesktopWidget()
-			height_min = 0.9*desktopWidget.height()
-			width_min = 0.96*desktopWidget.width()
+			height_min = int(0.9*desktopWidget.height())
+			width_min =int(0.96*desktopWidget.width())
 			self.setMinimumSize(width_min,height_min)
 		else:
 			self.setCentralWidget(self.centralWidget)
